backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Synqx Backend API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "synqx-core",
    "synqx-engine",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "structlog>=23.0.0",
    "alembic>=1.10.0",
    "pandas>=2.0.0",
    "python-jose>=3.3.0",
    "passlib>=1.7.0",
    "python-multipart>=0.0.6",
    "requests>=2.30.0",
    "google-genai>=1.59.0",
    "aiobotocore>=3.1.0",
    "aiohappyeyeballs>=2.6.1",
    "aiohttp>=3.13.3",
    "aioitertools>=0.13.0",
    "aiosignal>=1.4.0",
    "amqp>=5.3.1",
    "asn1crypto>=1.5.1",
    "attrs>=25.4.0",
    "azure-core>=1.38.0",
    "azure-storage-blob>=12.28.0",
    "bcrypt>=5.0.0",
    "billiard>=4.2.4",
    "black>=26.1.0",
    "boto3>=1.42.19",
    "botocore>=1.42.19",
    "cassandra-driver>=3.29.3",
    "celery>=5.6.2",
    "cffi>=2.0.0",
    "click-didyoumean>=0.3.1",
    "click-plugins>=1.1.1.2",
    "click-repl>=0.3.0",
    "cramjam>=2.11.0",
    "croniter>=6.0.0",
    "cryptography>=46.0.3",
    "databricks-sql-connector>=4.2.4",
    "deepdiff>=8.6.1",
    "dnspython>=2.8.0",
    "duckdb>=1.4.3",
    "elastic-transport>=9.2.1",
    "elasticsearch>=9.2.1",
    "email-validator>=2.3.0",
    "et-xmlfile>=2.0.0",
    "fastparquet>=2025.12.0",
    "filelock>=3.20.3",
    "flower>=2.0.1",
    "frozenlist>=1.8.0",
    "fsspec>=2026.1.0",
    "geomet>=1.1.0",
    "google-api-core>=2.29.0",
    "google-api-python-client>=2.188.0",
    "google-cloud-bigquery>=3.40.0",
    "google-cloud-core>=2.5.0",
    "google-cloud-storage>=3.8.0",
    "google-crc32c>=1.8.0",
    "google-resumable-media>=2.8.0",
    "googleapis-common-protos>=1.72.0",
    "grpcio>=1.76.0",
    "grpcio-status>=1.76.0",
    "httptools>=0.7.1",
    "humanize>=4.15.0",
    "inflection>=0.5.1",
    "iniconfig>=2.3.0",
    "invoke>=2.2.1",
    "isodate>=0.7.2",
    "jmespath>=1.0.1",
    "kafka-python-ng>=2.2.3",
    "kombu>=5.6.2",
    "librt>=0.7.8",
    "linkify-it-py>=2.0.3",
    "lxml>=6.0.2",
    "lz4>=4.4.5",
    "markdown-it-py>=4.0.0",
    "mdit-py-plugins>=0.5.0",
    "mdurl>=0.1.2",
    "more-itertools>=10.8.0",
    "multidict>=6.7.0",
    "mypy>=1.19.1",
    "mypy-extensions>=1.1.0",
    "oauthlib>=3.3.1",
    "openpyxl>=3.1.5",
    "oracledb>=3.4.1",
    "orderly-set>=5.5.0",
    "packaging>=25.0",
    "paramiko>=4.0.0",
    "pathspec>=1.0.3",
    "pika>=1.3.2",
    "platformdirs>=4.5.1",
    "pluggy>=1.6.0",
    "prometheus-client>=0.24.1",
    "prompt-toolkit>=3.0.52",
    "propcache>=0.4.1",
    "proto-plus>=1.27.0",
    "protobuf>=6.33.4",
    "psutil>=7.2.1",
    "psycopg2-binary>=2.9.11",
    "pyairtable>=3.3.0",
    "pyarrow>=23.0.0",
    "pybreaker>=1.4.1",
    "pycparser>=2.23",
    "pygments>=2.19.2",
    "pyjwt>=2.10.1",
    "pymongo>=4.16.0",
    "pymssql>=2.3.11",
    "pymysql>=1.1.2",
    "pynacl>=1.6.2",
    "pyopenssl>=25.3.0",
    "pyparsing>=3.3.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytokens>=0.4.0",
    "pyyaml>=6.0.3",
    "redis>=7.1.0",
    "requests-file>=3.0.1",
    "requests-toolbelt>=1.0.0",
    "rich>=14.2.0",
    "ruff>=0.14.13",
    "s3fs>=2026.1.0",
    "s3transfer>=0.16.0",
    "setuptools>=80.9.0",
    "shellingham>=1.5.4",
    "simple-salesforce>=1.12.9",
    "snowflake-connector-python>=4.2.0",
    "snowflake-sqlalchemy>=1.8.2",
    "sortedcontainers>=2.4.0",
    "textual>=7.3.0",
    "thrift>=0.20.0",
    "tomlkit>=0.14.0",
    "tornado>=6.5.4",
    "tqdm>=4.67.1",
    "typer>=0.21.1",
    "tzlocal>=5.3.1",
    "uc-micro-py>=1.0.3",
    "uritemplate>=4.2.0",
    "uvloop>=0.22.1",
    "vine>=5.1.0",
    "watchfiles>=1.1.1",
    "wcwidth>=0.2.14",
    "wrapt>=2.0.1",
    "yarl>=1.22.0",
    "zeep>=4.3.2",
]

[tool.uv.sources]
synqx-core = { path = "../libs/synqx-core", editable = true }
synqx-engine = { path = "../libs/synqx-engine", editable = true }

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "PL", "RUF"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true